pyproject.toml
[project]
name = "rlbind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "colabdesign @ https://github.com/sokrypton/ColabDesign.git",
  "jax>=0.7.2",
  "matplotlib>=3.10.7",
  "numpy>=2.3.4",
  "pip>=25.3",
  "polars>=1.35.2",
  "pydantic>=2.12.4",
  "pyrosetta-installer>=0.1.2",
  "requests>=2.32.5",
  "scipy>=1.16.3",
  "tdqm>=0.0.1",
  "umap-learn>=0.5.11",
]

[project.optional-dependencies]
rdkit = ["rdkit>=2025.9.4"]

[tool.uv.workspace]
members = ["bindcraft"]
